--- protocols/bftcosi/.gitignore
@@ -1,1 +1,0 @@
-block
